Default channel in channels stuff.
